Add tooltip-label example
