--- examples/tooltip-placement/index.tsx
@@ -1,20 +1,0 @@
-import * as Ariakit from "@ariakit/react";
-import "./style.css";
-
-export default function Example() {
-  const tooltip = Ariakit.useTooltipStore({ placement: "bottom-end" });
-  return (
-    <>
-      <Ariakit.TooltipAnchor
-        as={Ariakit.Button}
-        store={tooltip}
-        className="button"
-      >
-        Hover or focus on me
-      </Ariakit.TooltipAnchor>
-      <Ariakit.Tooltip store={tooltip} className="tooltip">
-        Tooltip
-      </Ariakit.Tooltip>
-    </>
-  );
-}
